+++ src/L1_Tuxedo.java
@@ -1,11 +1,11 @@
 import javax.swing.*;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class L1_Tuxedo extends Cat{
     L1_Tuxedo() {
         this.kitten = new ImageIcon("res/kitten1HD.png");
         this.levelname = "Level 1: Tuxedo Cat";
         this.level = 1;
-        this.nextlevel = BigInteger.valueOf(1000000);
+        this.nextlevel = BigDecimal.valueOf(1000000);
     }
 }
