+++ src/L3_Tortoiseshell.java
@@ -1,11 +1,11 @@
 import javax.swing.*;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class L3_Tortoiseshell extends Cat{
     L3_Tortoiseshell() {
         this.kitten = new ImageIcon("res/kitten3HD.png");
         this.levelname = "Level 3: Tortoiseshell Cat";
         this.level = 3;
-        this.nextlevel = new BigInteger("50000000000000");
+        this.nextlevel = new BigDecimal("50000000000000");
     }
 }
