+++ src/L4_Orange.java
@@ -1,11 +1,11 @@
 import javax.swing.*;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class L4_Orange extends Cat{
     L4_Orange() {
         this.kitten = new ImageIcon("res/kitten4HD.png");
         this.levelname = "Level 4: Orange Cat";
         this.level = 4;
-        this.nextlevel = new BigInteger("2000000000000000");
+        this.nextlevel = new BigDecimal("2000000000000000");
     }
 }
